// src/app/open-props/lib/colors.stylex.d.ts
type TColors = Readonly<{
  gray0: string;
  gray1: string;
  gray2: string;
  gray3: string;
  gray4: string;
  gray5: string;
  gray6: string;
  gray7: string;
  gray8: string;
  gray9: string;
  gray10: string;
  gray11: string;
  gray12: string;
  stone0: string;
  stone1: string;
  stone2: string;
  stone3: string;
  stone4: string;
  stone5: string;
  stone6: string;
  stone7: string;
  stone8: string;
  stone9: string;
  stone10: string;
  stone11: string;
  stone12: string;
  red0: string;
  red1: string;
  red2: string;
  red3: string;
  red4: string;
  red5: string;
  red6: string;
  red7: string;
  red8: string;
  red9: string;
  red10: string;
  red11: string;
  red12: string;
  pink0: string;
  pink1: string;
  pink2: string;
  pink3: string;
  pink4: string;
  pink5: string;
  pink6: string;
  pink7: string;
  pink8: string;
  pink9: string;
  pink10: string;
  pink11: string;
  pink12: string;
  purple0: string;
  purple1: string;
  purple2: string;
  purple3: string;
  purple4: string;
  purple5: string;
  purple6: string;
  purple7: string;
  purple8: string;
  purple9: string;
  purple10: string;
  purple11: string;
  purple12: string;
  violet0: string;
  violet1: string;
  violet2: string;
  violet3: string;
  violet4: string;
  violet5: string;
  violet6: string;
  violet7: string;
  violet8: string;
  violet9: string;
  violet10: string;
  violet11: string;
  violet12: string;
  indigo0: string;
  indigo1: string;
  indigo2: string;
  indigo3: string;
  indigo4: string;
  indigo5: string;
  indigo6: string;
  indigo7: string;
  indigo8: string;
  indigo9: string;
  indigo10: string;
  indigo11: string;
  indigo12: string;
  blue0: string;
  blue1: string;
  blue2: string;
  blue3: string;
  blue4: string;
  blue5: string;
  blue6: string;
  blue7: string;
  blue8: string;
  blue9: string;
  blue10: string;
  blue11: string;
  blue12: string;
  cyan0: string;
  cyan1: string;
  cyan2: string;
  cyan3: string;
  cyan4: string;
  cyan5: string;
  cyan6: string;
  cyan7: string;
  cyan8: string;
  cyan9: string;
  cyan10: string;
  cyan11: string;
  cyan12: string;
  teal0: string;
  teal1: string;
  teal2: string;
  teal3: string;
  teal4: string;
  teal5: string;
  teal6: string;
  teal7: string;
  teal8: string;
  teal9: string;
  teal10: string;
  teal11: string;
  teal12: string;
  green0: string;
  green1: string;
  green2: string;
  green3: string;
  green4: string;
  green5: string;
  green6: string;
  green7: string;
  green8: string;
  green9: string;
  green10: string;
  green11: string;
  green12: string;
  lime0: string;
  lime1: string;
  lime2: string;
  lime3: string;
  lime4: string;
  lime5: string;
  lime6: string;
  lime7: string;
  lime8: string;
  lime9: string;
  lime10: string;
  lime11: string;
  lime12: string;
  yellow0: string;
  yellow1: string;
  yellow2: string;
  yellow3: string;
  yellow4: string;
  yellow5: string;
  yellow6: string;
  yellow7: string;
  yellow8: string;
  yellow9: string;
  yellow10: string;
  yellow11: string;
  yellow12: string;
  orange0: string;
  orange1: string;
  orange2: string;
  orange3: string;
  orange4: string;
  orange5: string;
  orange6: string;
  orange7: string;
  orange8: string;
  orange9: string;
  orange10: string;
  orange11: string;
  orange12: string;
  choco0: string;
  choco1: string;
  choco2: string;
  choco3: string;
  choco4: string;
  choco5: string;
  choco6: string;
  choco7: string;
  choco8: string;
  choco9: string;
  choco10: string;
  choco11: string;
  choco12: string;
  brown0: string;
  brown1: string;
  brown2: string;
  brown3: string;
  brown4: string;
  brown5: string;
  brown6: string;
  brown7: string;
  brown8: string;
  brown9: string;
  brown10: string;
  brown11: string;
  brown12: string;
  sand0: string;
  sand1: string;
  sand2: string;
  sand3: string;
  sand4: string;
  sand5: string;
  sand6: string;
  sand7: string;
  sand8: string;
  sand9: string;
  sand10: string;
  sand11: string;
  sand12: string;
  camo0: string;
  camo1: string;
  camo2: string;
  camo3: string;
  camo4: string;
  camo5: string;
  camo6: string;
  camo7: string;
  camo8: string;
  camo9: string;
  camo10: string;
  camo11: string;
  camo12: string;
  jungle0: string;
  jungle1: string;
  jungle2: string;
  jungle3: string;
  jungle4: string;
  jungle5: string;
  jungle6: string;
  jungle7: string;
  jungle8: string;
  jungle9: string;
  jungle10: string;
  jungle11: string;
  jungle12: string;
}>;
export declare const colors: VarGroup<TColors>;